[Decomp][Player] Cohesion split: extract mirror-timer system to PlayerMirror.cpp
Moves the contiguous mirror-timer + environmental-damage + drunk
state block out of Player.cpp into a new file under src/game/Object/.

Player.cpp: 13,856 -> 13,465 lines (-391).
PlayerMirror.cpp: 474 lines, 8 methods.

Methods moved:
- SendMirrorTimer, StopMirrorTimer
- EnvironmentalDamage
- getMaxTimer, UpdateMirrorTimers, HandleDrowning
- HandleSobering, SetDrunkValue

No header change; no signature change; no behaviour change.
